Remove support for serializing longs
Longs are of different size on Windows and Linux, which means
serializing longs leads to non-portable savegames or game data.

Also, the serializing of longs was via doubles, which means we couldn't
support longs with full precision anyway.

All the things that were serialized as longs are now done as ints:
- Charges of various sorts.
- Cash amounts for the player and mission rewards.
- Map symbols (which are usually unicode codepoints, or a handful of
  other special values).

+++ src/itype.h
@@ -114,7 +114,7 @@ struct islot_comestible {
     std::string tool = "null";
 
     /** Defaults # of charges (drugs, loaf of bread? etc) */
-    long def_charges = 1;
+    int def_charges = 1;
 
     /** effect on character thirst (may be negative) */
     int quench = 0;
@@ -638,7 +638,7 @@ struct islot_ammo : common_ranged_data {
     /**
      * Default charges.
      */
-    long def_charges = 1;
+    int def_charges = 1;
 
     /**
      * TODO: document me.
